Removed speed index everywhere and using floats now; New textures 5..140 with step 5.

+++ TLM/TLM/State/Configuration.cs
@@ -12,9 +12,9 @@ namespace TrafficManager {
 		[Serializable]
 		public class LaneSpeedLimit {
 			public uint laneId;
-			public ushort speedLimit;
+			public float speedLimit;
 
-			public LaneSpeedLimit(uint laneId, ushort speedLimit) {
+			public LaneSpeedLimit(uint laneId, float speedLimit) {
 				this.laneId = laneId;
 				this.speedLimit = speedLimit;
 			}
